+++ exporters/otlp/src/otlp_grpc_log_record_exporter.cc
@@ -134,18 +134,24 @@ opentelemetry::sdk::common::ExportResult OtlpGrpcLogRecordExporter::Export(
   return sdk::common::ExportResult::kSuccess;
 }
 
-bool OtlpGrpcLogRecordExporter::Shutdown(std::chrono::microseconds /* timeout */) noexcept
+bool OtlpGrpcLogRecordExporter::Shutdown(std::chrono::microseconds timeout) noexcept
 {
   const std::lock_guard<opentelemetry::common::SpinLockMutex> locked(lock_);
   is_shutdown_ = true;
+#ifdef ENABLE_ASYNC_EXPORT
+  return client_->Shutdown(timeout);
+#else
   return true;
+#endif
 }
 
-bool OtlpGrpcLogRecordExporter::ForceFlush(std::chrono::microseconds /* timeout */) noexcept
+bool OtlpGrpcLogRecordExporter::ForceFlush(std::chrono::microseconds timeout) noexcept
 {
-  // TODO: When we implement async exporting in OTLP gRPC exporter in the future, we need wait the
-  //       running exporting finished here.
+#ifdef ENABLE_ASYNC_EXPORT
+  return client_->ForceFlush(timeout);
+#else
   return true;
+#endif
 }
 
 bool OtlpGrpcLogRecordExporter::isShutdown() const noexcept