+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Increment the:
   [#2398](https://github.com/open-telemetry/opentelemetry-cpp/pull/2398)
 * [EXPORTER] Rework OTLP/HTTP and OTLP/GRPC exporter options
   [#2388](https://github.com/open-telemetry/opentelemetry-cpp/pull/2388)
+* [EXPORTER] Add async exporting for OTLP/GRPC exporter
+  [#2407](https://github.com/open-telemetry/opentelemetry-cpp/pull/2407)
 
 Important changes: